requirements.txt
aiogram>=3.0
sqlalchemy>=2.0
alembic
asyncpg
psycopg2-binary
pgvector
pandas
openpyxl
google-generativeai
gigachat
aiohttp
python-dotenv
pytest
pytest-asyncio
pytest-cov
pytest-mock
aiogram[test]
aiosqlite
